+++ demo/DemoV3/DemoV3Functions.generated.cs
@@ -13,6 +13,7 @@ using Microsoft.Extensions.Logging;
 
 using Microsoft.Azure.WebJobs.Extensions.DurableTask;
 using Newtonsoft.Json.Linq;
+using DemoV3.Model.Base;
 using DemoV3.Model;
 
 namespace DemoV3