+++ demo/DemoV3/IDemoV3Service.generated.cs
@@ -6,6 +6,7 @@ using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Azure.WebJobs.Extensions.DurableTask;
 using Newtonsoft.Json.Linq;
+using DemoV3.Model.Base;
 using DemoV3.Model;
 
 namespace DemoV3