+++ demo/DemoV3/Model/Base/Address.generated.cs
@@ -4,7 +4,7 @@ using Newtonsoft.Json.Linq;
 using System.Collections.Generic;
 using System.Linq;
 
-namespace DemoV3.Model {
+namespace DemoV3.Model.Base {
     /// AUTOGENERED BY caffoa ///
     [JsonObject(MemberSerialization.OptIn)]
     public partial class Address {