+++ demo/DemoV3/Model/GuestUser.generated.cs
@@ -3,6 +3,7 @@ using Newtonsoft.Json;
 using Newtonsoft.Json.Linq;
 using System.Collections.Immutable;
 using System.Linq;
+using DemoV3.Model.Base;
 
 namespace DemoV3.Model {
     /// AUTOGENERED BY caffoa ///