+++ demo/DemoV3/Model/User.generated.cs
@@ -5,6 +5,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Collections.Immutable;
 using Caffoa.JsonConverter;
+using DemoV3.Model.Base;
 
 namespace DemoV3.Model {
     /// AUTOGENERED BY caffoa ///