+++ changelog.md
@@ -1,5 +1,8 @@
 # caffoa changelog
 
+### 1.2.0
+* new option: `splitByTag`. When this config option is set to true, multiple interfaces and function classes will be generated based on the first tag of each path
+
 ### 1.1.0
 * new option: use `parsePathParameters` to let caffoa handle the path parameter parsing instead of azure functions.
 Additionally, you can individualize the parsing, by implementing your own logic.